requirements.txt
ccxt==4.5.0
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
PyYAML==6.0.2
